requirements.txt
flask-discoverer
requests